freetype: replace corrupted makefiles
the only difference of the upstream files to the lib-alfont version was
that 3 makefile snippets were cut off in the middle.
the originals contain bogus \xff characters, however it seems make is able
to handle them just like it would handle a space character.
this must have caused the corruption in the lib-alfont version.
i now replace the \xff characters with spaces in builds/detect.mk and
builds/win32/detect.mk, which makes it possible to run configure successfully.
the ags version, due to being cut off, fails immediately with a syntax error.

the builds/mac/freetype.make file still contains the original \xff, as i don't
know how to fix it. the file seems to use a different buildsystem i'm unfamiliar
with. also, it's only for powerpc macs which are very rare these days, so this
code is likely unused and will stay so.